Rename `lib_zarr` into `lib_write`
