+++ tests/tasks/test_workflows_multiplexing.py
@@ -19,7 +19,7 @@ from devtools import debug
 
 from ._validation import check_file_number
 from ._validation import validate_schema
-from fractal_tasks_core.lib_zarr import OverwriteNotAllowedError
+from fractal_tasks_core.lib_write import OverwriteNotAllowedError
 from fractal_tasks_core.tasks.copy_ome_zarr import (
     copy_ome_zarr,
 )
