+++ tests/tasks/test_workflows_napari_workflows.py
@@ -27,7 +27,7 @@ from ._validation import validate_schema
 from .lib_empty_ROI_table import _add_empty_ROI_table
 from fractal_tasks_core.lib_input_models import NapariWorkflowsInput
 from fractal_tasks_core.lib_input_models import NapariWorkflowsOutput
-from fractal_tasks_core.lib_zarr import OverwriteNotAllowedError
+from fractal_tasks_core.lib_write import OverwriteNotAllowedError
 from fractal_tasks_core.tasks.napari_workflows_wrapper import (
     napari_workflows_wrapper,
 )
