+++ tests/test_unit_zarr.py
@@ -4,11 +4,11 @@ import pytest
 import zarr
 from devtools import debug
 
-from fractal_tasks_core.lib_zarr import _write_elem_with_overwrite
-from fractal_tasks_core.lib_zarr import open_zarr_group_with_overwrite
-from fractal_tasks_core.lib_zarr import OverwriteNotAllowedError
-from fractal_tasks_core.lib_zarr import prepare_label_group
-from fractal_tasks_core.lib_zarr import write_table
+from fractal_tasks_core.lib_write import _write_elem_with_overwrite
+from fractal_tasks_core.lib_write import open_zarr_group_with_overwrite
+from fractal_tasks_core.lib_write import OverwriteNotAllowedError
+from fractal_tasks_core.lib_write import prepare_label_group
+from fractal_tasks_core.lib_write import write_table
 
 
 def test_open_zarr_group_with_overwrite(tmp_path, caplog):
